+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+ * Reintroduction of the interactive launch of parallel MPI programs (#661 and !616)
  * Add a timeout to makedepf90 to prevent stalling (#671 and !617)
  * Avoid calling `printf` in `tools/midas_scripts/midas.reunir_obs_mpi`
    with potentially very long arguments (#654 and !609)